fixing warnings refs #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@10167 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/geom/Position2DVector.cpp
@@ -1021,7 +1021,7 @@ Position2DVector::removeColinearPoints() {
 bool 
 Position2DVector::operator==(const Position2DVector &v2) const {
     if (size() == v2.size()) {
-        for (size_t i = 0; i < size(); i++) {
+        for (int i = 0; i < (int)size(); i++) {
             if ((*this)[i] != v2[i]) {
                 return false;
             }
